renamed quickactions to index and fixed imports
